Move performance track property serialization to shared
